code share and test apply upgrade logic
